event: restrict visibility of backend functions and consolidate tests
